move Phase1Strip from SimplePixelTopology.h to SimplePixelStripTopology.h
