+++ RecoLocalTracker/SiPixelRecHits/src/PixelCPEFastParams.cc
@@ -1,6 +1,6 @@
 #include "FWCore/Utilities/interface/typelookup.h"
 #include "RecoLocalTracker/SiPixelRecHits/interface/PixelCPEFastParamsDevice.h"
-#include "Geometry/CommonTopologies/interface/SimplePixelTopology.h"
+#include "Geometry/CommonTopologies/interface/SimplePixelStripTopology.h"
 
 using PixelCPEFastParamsPhase1 = PixelCPEFastParamsDevice<alpaka_common::DevHost, pixelTopology::Phase1>;
 using PixelCPEFastParamsPhase2 = PixelCPEFastParamsDevice<alpaka_common::DevHost, pixelTopology::Phase2>;
