+++ RecoTracker/PixelVertexFinding/plugins/alpaka/vertexFinder.dev.cc
@@ -1,6 +1,7 @@
 #include <alpaka/alpaka.hpp>
 #include "DataFormats/TrackSoA/interface/alpaka/TrackUtilities.h"
 #include "DataFormats/Vertex/interface/alpaka/ZVertexUtilities.h"
+#include "Geometry/CommonTopologies/interface/SimplePixelStripTopology.h"
 #include "HeterogeneousCore/AlpakaInterface/interface/workdivision.h"
 #include "HeterogeneousCore/AlpakaInterface/interface/traits.h"
 #include "PixelVertexWorkSpaceUtilitiesAlpaka.h"
